% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_forclass.R
\name{check_forclass}
\alias{check_forclass}
\title{Check Classification Models}
\usage{
check_forclass(models)
}
\arguments{
\item{models}{A character vector of model names to check.}
}
\value{
A data frame containing model details where \code{forClass} is \code{TRUE}.
}
\description{
This function verifies if the given machine learning models support
classification tasks by ensuring the \code{forClass} column is \code{TRUE}
in their model lookup details.
}
\examples{
check_forclass(c("lvq", "gbm", "svmRadial"))
}
