% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparer.R
\name{compare_models}
\alias{compare_models}
\title{Compare Classification Models}
\usage{
compare_models(data, target_var, train_prop = 0.8, seed = 3456)
}
\arguments{
\item{data}{A data frame containing the dataset to be used for modeling.}

\item{target_var}{The name of the target variable in the dataset.}

\item{train_prop}{The proportion of data to be used for training (default
is 0.8).}

\item{seed}{The random seed for reproducibility (default is 3456).}
}
\value{
A list containing the trained models and their performance metrics.
}
\description{
This function compares 13 different classification models using the caret
package.
}
\examples{
library(mlbench)
data("PimaIndiansDiabetes", package = "mlbench")
results <- compare_models(PimaIndiansDiabetes, "diabetes")
}
