% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparer.R
\name{model_comparer}
\alias{model_comparer}
\title{Compare Classification Models}
\usage{
model_comparer(
  data,
  target_var,
  train_prop = 0.8,
  seed = 3456,
  for_utest = FALSE
)
}
\arguments{
\item{data}{A data frame containing the dataset to be used for modeling.}

\item{target_var}{The name of the target variable in the dataset.}

\item{train_prop}{The proportion of data to be used for training (default
is 0.8).}

\item{seed}{The random seed for reproducibility (default is 3456).}

\item{for_utest}{only for unit test when is TRUE (FALSE by default).}
}
\value{
A list containing the trained models and their performance metrics.
}
\description{
This function compares 13 different classification models using the caret
package.
}
\examples{
library(mlbench)
data("PimaIndiansDiabetes", package = "mlbench")
#results <- model_comparer(PimaIndiansDiabetes, "diabetes")
}
