% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCMC-package.R
\docType{package}
\name{TCMC-package}
\alias{TCMC}
\alias{TCMC-package}
\title{TCMC: Compare Classification Models}
\description{
Streamlined approach to train, evaluate, and compare different classification algorithms, by offering the user a robust framework for model selection in machine learning projects.
}
\seealso{
Useful links:
\itemize{
  \item \url{}
  \item Report bugs at \url{https://github.com/danymukesha/TCMC/issues}
}

}
\author{
\strong{Maintainer}: Dany Mukesha \email{danymukesha@gmail.com} (\href{https://orcid.org/0009-0001-9514-751X}{ORCID})

}
\keyword{internal}
