% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparer.R
\name{plot_importance}
\alias{plot_importance}
\title{Plot Variable Importance}
\usage{
plot_importance(model, model_name)
}
\arguments{
\item{model}{A trained model object.}

\item{model_name}{The name of the model (used for the plot title).}
}
\value{
A ggplot object showing variable importance.
}
\description{
This function plots the variable importance for a given model.
}
\examples{
library(mlbench)
data("PimaIndiansDiabetes", package = "mlbench")
#results <- model_comparer(PimaIndiansDiabetes, "diabetes")
#plot_importance(results$trained_models$lvq, "LVQ")
}
